api-server/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
aiosqlite>=0.20,<1.0
httpx>=0.27,<1.0
PyYAML>=6.0,<7.0
pydantic-settings>=2.3,<3.0
asyncpg>=0.29,<1.0

